Make Cinemeta client replacable
- Introduces Metafetcher interface
- Cinemeta is still used by default, but can be replaced

The returned meta type is from the cinemeta package, but this might
change in the future to make go-stremio completely independent from the
cinemeta package.

Note: The main reason for this change is to be able to use
imdb2meta (https://github.com/Deflix-tv/imdb2meta), which is faster,
more up-to-date and makes us independent from Cinemeta.